Move pictures into games
